requirements.txt
numpy
pydicom
matplotlib
pillow
scipy